feat: adding functions login, logout and restriction.

+++ base/views.py
@@ -1,13 +1,43 @@
-from .models import Room, Topic
-from .forms import RoomForm
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.models import User
 from django.db.models import Q
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 
-rooms = [
-    {'id': 1, 'name': 'Lets learn python!'},
-    {'id': 2, 'name': 'Lets learn Vue!'},
-    {'id': 3, 'name': 'Lets learn Flask!'},
-]
+from .forms import RoomForm
+from .models import Room, Topic
+
+
+def login_page(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist.')
+
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request=request, user=user)
+            return redirect('home')
+        else:
+            messages.error(request=request,
+                           message='Username or password does not exist.')
+    context = {}
+    return render(request=request, template_name='base/login_register.html',
+                  context=context)
+
+
+def logout_page(request):
+    logout(request=request)
+    return redirect('login')
 
 
 def home(request):
@@ -29,6 +59,7 @@ def room(request, pk):
     return render(request=request, template_name='base/room.html', context=context)
 
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -40,9 +71,13 @@ def create_room(request):
     return render(request=request, template_name='base/room_form.html', context=context)
 
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!!')
 
     if request.method == 'POST':
         form = RoomForm(data=request.POST, instance=room)
@@ -53,8 +88,11 @@ def update_room(request, pk):
     return render(request=request, template_name='base/room_form.html', context=context)
 
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here!!!')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
